removes extra auth-receiver file
