+++ auth-receiver.js
@@ -1,15 +1,18 @@
+// process.env.NODE_TLS_REJECT_UNAUTHORIZED = "0";
 const Imap = require('imap');
 const inspect = require('util').inspect;
-require('dotenv').config(); 
 
-module.exports = function (servo) {
-
-    var rightnow = new Date().getTime()
+module.exports = function (servo, io) {
+    io.on('connection', function(socket){
+    var state = {
+        rightnow: new Date(),
+        mailTick: 0
+    }
 
     //links app to email and detects events
     let imap = new Imap({
-        user: `${process.env.emailAdd}`,
-        password: `${process.env.emailPW}`,
+        user: 'gtbc2018facebell@gmail.com',
+        password: '!2018facebell',
         host: 'imap.gmail.com',
         port: 993,
         tls: true
@@ -22,15 +25,18 @@ module.exports = function (servo) {
 
     //event fire on new mail, sends status to servo function
     imap.on('mail', function (mail) {
+        console.log('--------------> MAIL EVENT WAS FIRED <------------');
         let authKey;
         openInbox(function (err, box) {
             if (err) throw err;
             var f = imap.seq.fetch(box.messages.total + ':*', { bodies: ['HEADER.FIELDS (FROM)', 'TEXT'] });
             f.on('message', function (msg, seqno) {
+                console.log('Message #%d', seqno);
                 var prefix = '(#' + seqno + ') ';
                 msg.on('body', function (stream, info) {
 
                     if (info.which === 'TEXT')
+                        console.log(prefix + 'Body [%s] found, %d total bytes', inspect(info.which), info.size);
                     var buffer = '', count = 0;
                     stream.on('data', function (chunk) {
                         count += chunk.length;
@@ -40,10 +46,19 @@ module.exports = function (servo) {
                         authKey = Object.values(Imap.parseHeader(buffer))[0]
                         if (authKey !== undefined) {
                             authKey = authKey[0];
+                            console.log(authKey);
                         }
 
                         if (authKey === `${process.env.phoneNum}`) {
+                            
                             servo.max();
+                        
+                        
+                                    console.log("emit-unlock sent");
+                        
+                                    io.emit('emit-unlock', { success : true});
+                                
+                            
                             setTimeout(() => {
                                 servo.center();
                             }, 1500);
@@ -56,16 +71,18 @@ module.exports = function (servo) {
     });
 
 
-    //more boilerplate, searches mail on start and ive changed code to not disconnect after fetch to keep listener on
+    //more boilerplate, searches mail on start and ive changed code to not disconnect after fetch to keep listener on and not write txt files to root
     imap.once('ready', function () {
         openInbox(function (err, box) {
             if (err) throw err;
-            imap.search(['UNSEEN', ['SENTSINCE', rightnow]], function (err, results) {
+            imap.search(['DELETED', ['SENTSINCE', state.rightnow.getTime()]], function (err, results) {
                 if (err) throw err;
                 imap.on('message', function (msg, seqno) {
                     msg.on('body', function (stream, info) {
+                        console.log(prefix + 'Body');
                     });
                     msg.once('attributes', function (attrs) {
+                        console.log(prefix + 'Attributes: % s', inspect(attrs, false, 8));
                     });
                 });
             });
@@ -78,4 +95,5 @@ module.exports = function (servo) {
         console.log('Connection ended');
     });
     imap.connect();
+})
 }
